feat: support debugging with `--manual` inside browser
Currently when this jasmine test framework is used, when manually
connecting to a browser, there is zero log output about failing, or
succeeding tests. This is now fixed with this change where we properly
embed the Jasmine console reporter into the browser; rendering nice
and useful output; similar to the real console output.